main: Move renderOrHydrateApp to dedicated file
